+++ src/store_queue.h
@@ -63,7 +63,7 @@ class StoreQueue {
 
   // WARNING: don't expect this to be exact, because it could change after you check.
   //          This is only for hueristics to decide when we're overloaded.
-  unsigned long getSize(bool lock = true);
+  unsigned long long getSize(bool lock = true);
  private:
   void storeInitCommon();
   void configureInline(pStoreConf configuration);
@@ -94,7 +94,7 @@ class StoreQueue {
   cmd_queue_t cmdQueue;
   boost::shared_ptr<logentry_vector_t> msgQueue;
   boost::shared_ptr<logentry_vector_t> failedMessages;
-  unsigned long msgQueueSize;   // in bytes
+  unsigned long long msgQueueSize;   // in bytes
   pthread_t storeThread;
 
   // Mutexes
@@ -112,11 +112,11 @@ class StoreQueue {
   bool multiCategory; // Whether multiple categories are handled
 
   // configuration
-  std::string   categoryHandled;  // what category this store is handling
-  time_t        checkPeriod;      // how often to call periodicCheck in seconds
-  unsigned long targetWriteSize;  // in bytes
-  time_t        maxWriteInterval; // in seconds
-  bool          mustSucceed;      // Always retry even if secondary fails
+  std::string        categoryHandled;  // what category this store is handling
+  time_t             checkPeriod;      // how often to call periodicCheck in seconds
+  unsigned long long targetWriteSize;  // in bytes
+  time_t             maxWriteInterval; // in seconds
+  bool               mustSucceed;      // Always retry even if secondary fails
 
   // Store that will handle messages. This can contain other stores.
   boost::shared_ptr<Store> store;
